feat(presenters): generalize chart presenter to timeline tracker for extendability
