+++ src/presenters/index.ts
@@ -4,8 +4,12 @@ import {
 import {
   ChartPresenter,
 } from './chart-presenter/index.js';
+import {
+  TimelineDataPresenter,
+} from './timeline-data-presenter/index.js';
 
 export const nativePresenters = {
   jsonChunkPresenter: JsonChunkPresenter,
   chartPresenter: ChartPresenter,
+  timelineDataPresenter: TimelineDataPresenter,
 } as const;
