+++ src/presenters/timeline-data-presenter/index.ts
@@ -1,0 +1,2 @@
+export * from './timeline-data-presenter.js';
+export type * from './types.js';
